Issue 93 (#4)
* Add 'reverse_mapping_fields' to registry. This is a simple reversal
of the 'mapping_fields' dictionary keys and values. The purpose of
this dictionary is so that the LogEntry model's 'changes_display_dict'
method can reverse a human readable field name returned from 'changes'
so that it can do extra checks on that field. Without this we would get
a FieldDoesNotExist error when trying to lookup fields on the model in
the 'changes_display_dict' method so that we could introspect attributes
about them when determining how to display human readable changes.